requirements.txt
openai == 1.99.9
pillow == 11.3.0
requests == 2.32.4
